# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Paws and Homes API
  version: 1.0.0
  description: Backend API for the Paws and Homes pet adoption portal
servers:
  - url: /
paths:
...
  /api/pets:
    get:
      x-wso2-disable-security: true
      summary: Retrieve all pets
      responses:
        "200":
          description: Pet list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"

    post:
      summary: Create a new pet
      responses:
        "201":
          description: Pet created successfully
        "401":
          description: Authentication required
        "403":
          description: Shelter Staff role required
...
  /api/applications:
    post:
      summary: Create a new application
      responses:
        "201":
          description: Application created successfully
        "401":
          description: Authentication required
...
components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          oneOf:
            - type: integer
            - type: string
          example: 1
        name:
          type: string
          example: Buddy
        species:
          type: string
          example: Dog
        breed:
          type: string
          example: Golden Retriever
        age:
          type: integer
          example: 3
        status:
          type: string
          example: Available
        createdBy:
          type: string
          example: shelter.staff@example.com